Relax type requirement of EditorScene.viewport property
The goal of this patch series is to make editorlib compatible with Qt 5.14+.
Currently the viewport freezes because of some changes made between Qt 5.12
and 5.14. I suspect it would be synchronization related.

Instead of backporting changes to EditorViewport, I'm going to take the easy
path, which is to replace EditorViewport with Scene3D. That's why the viewport
type is relaxed to QQuickItem. Both EditorScene and EditorCameraController
only needs the viewport geometry, so QQuickItem type should be enough.

Change-Id: Ia28e82dc4c7c0caff9206002a29c98c2b8b52fb9