Switch viewport to Scene3D
Otherwise the viewport would deadlock on Qt 5.14 at:
  Qt3DRender::QRenderAspectPrivate::get(m_renderAspect))->renderSynchronous()
  FrameBufferObjectRenderer::render()

The most weird part of this patch is the ownership of the root entity. If it
were passed directly in to Scene3D, the entire entity tree would be deleted
along with Scene3D and the destructor of EditorScene would crash because of
double free. Before, this problem was invisible probably because EditorScene
would be deleted slightly earlier than the QAspectEngine.

hoverEnabled is set to true as it was set as such in EditorViewportItem,
but I don't know if it was working. There is a hover-enabled 2D MouseArea
covering the surface of the viewport.

Apparently on Qt 5.14.2, QObjectPicker of the helper arrows isn't working
until new object is added. This problem doesn't occur on Qt 5.12, so I think
it is unrelated issue.

Change-Id: I6d879eed4c634fa8cb639a2018e6fff71dfab158